Consolidate the useCreateWorkspace suites and close residual offer races
Two follow-ups to the first-run-offer settling:

useCreateWorkspace.test.ts and useCreateWorkspace.test.tsx share a module
stem, and TypeScript's project service resolves them to one module — the
.tsx twin then fails type-aware linting with "file was not found in any of
the provided project(s)" (masked locally by the eslint cache). Fold the
agent-type gating suite into the .tsx file and drop the .ts twin.

The offer's open state is global (the dialog is hosted in AppShell), so an
offer that fires while a route change is in flight survives it and lands on
top of the destination page:

- navigate_to_settings_page now also requires a clear overlay after landing
  on Settings; the retry body's Escape dismisses one that rode through.
- navigate_to_home_page dismisses an open dialog before clicking, mirroring
  navigate_to_workspace — with the modal up the sidebar is never clickable,
  so dismissal is always the precondition.
- test_workspace_deleted_via_cli_disappears_from_ui settles the offer that
  deterministically re-fires when the CLI delete empties the list while the
  page is parked on Home.
- test_copy_branch_name opens the new-workspace form explicitly instead of
  relying on a reload to re-offer it; the form re-sources the repo's
  branches on every open, so the reload was unnecessary anyway.
